First iteration of calculator
Note: This is my first attempt at using extended git commit messages, so hopefully this is legible.
This is my first iteration of the calculator portion of RB101. In an effort to challenge myself, I opted to design the program myself before going through the Walk-through portion of the lesson, so it might be missing features that the lesson adds.
What it CAN do:
perform the following operations on two integers: add, subtract, multiply, divide, modulo, exponent
Opt to store the results in an array as "memory", allowing the user to recall their results to perform multiple calculations with the results of these basic calculations.
What it CANT do:
Perform complex operations using three or more values.
Perform calculations using float values.
add, remove, or otherwise modify values stored in the memory array.