Crazy RPS!
I went a bit crazy with the rock paper scissors, but I wanted to challenge myself a bit so my game can take as many options for "weapons" as you can think of, as long as it is an odd number of options.
All you have to do is edit the choice_doc.yml, add in pairs of "weapons", as many as you want, and the game will run smoothly. Currently has 27 weapons and works, tested around 50 times.
The table code is not my own, but that wasn't my focus for this task.
Code is a bit all over the place, still not great at creating truly abstract methods, and organizing methods by what they do, but I wanted to move on as I feel I've gone a bit out of scope of these lessons anyways.